% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFcheck.R
\name{BFcheck}
\alias{BFcheck}
\title{Brute-Force Check of Probability Matrix}
\usage{
BFcheck(P, l = 1e-04, u = 0.9999)
}
\arguments{
\item{P}{Estimated probability matrix.}

\item{l}{Lower bound of the probability matrix, i.e. set all entries that not greater than 0 to \code{l}. 0.0001 by default.}

\item{u}{Upper bound of the probability matrix, i.e. set all entries that not less than 1 to be \code{u}. 0.9999 by default.}
}
\value{
A matrix of which all entries are between 0 and 1.
}
\description{
Check whether all entries of the estimated probabiltiy matrix are between 0 and 1.
Set the ones that not greater than 0 to be a lower bound and the ones that not less than 1 to be a upper bound.
}
\examples{
P <- matrix(runif(25,-1,2), nrow = 5)
BFcheck(P)

}
\seealso{
\code{\link{grdpg}}
}
\author{
Cong Mu \email{placid8889@gmail.com}
}
