% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cov.R
\name{compute_cov}
\alias{compute_cov}
\title{Intermediate Function}
\usage{
compute_cov(i, l1, l2, K, meanE, muhats, deltainv, eta, theta, zeta, Ipq)
}
\description{
Intermediate function.
}
\seealso{
\code{\link{grdpg}}
}
\author{
Cong Mu \email{placid8889@gmail.com}
}
