% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimselect.R
\name{dimselect}
\alias{dimselect}
\title{Dimension Selection}
\usage{
dimselect(s, n = 3)
}
\arguments{
\item{s}{A vector of ordered singular values.}

\item{n}{Number of elbows to be returned. 3 by default.}
}
\value{
A list containing the following:
\describe{
\item{\code{value}}{The singular value associated with the elbow.}
\item{\code{elbow}}{The embeded dimension (index of the elbow).}
}
}
\description{
Select embeded dimension of latent position by finding the \sQuote{elbow} of the scree plot using \href{http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.90.3768&rep=rep1&type=pdf}{profile likelihood}.
}
\references{
\code{\link[graphstats]{gs.dim.select}}
}
\seealso{
\code{\link{grdpg}}
}
