% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatebeta.R
\name{estimatebeta}
\alias{estimatebeta}
\title{Estimate beta}
\usage{
estimatebeta(BXhat, cov, covariates_block)
}
\arguments{
\item{BXhat}{Estimated block probability matrix. Should be a \code{k} by \code{k} matrix where \code{k} is the number of blocks.}

\item{cov}{A vector specifying possible value that each covariate could take. For example, if two binary covariates, then \code{cov <- c(2, 2)}.}

\item{covariates_block}{Estimated covariates for each block. Should be a \code{k} by \code{c} matrix or dataframe where \code{k} is the number of blocks and \code{c} is the number of covariates.}
}
\value{
A list containing all estimated beta. Each element of the list, i.e. the length of the list equals to the number of covariates.
}
\description{
Estimate beta (the effect of covariates).
}
\seealso{
\code{\link{grdpg}}
}
\author{
Cong Mu \email{placid8889@gmail.com}
}
