% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatebeta2.R
\name{estimatebeta2}
\alias{estimatebeta2}
\title{Estimate beta}
\usage{
estimatebeta2(BXhat, cov, covariates, clusters_cov)
}
\arguments{
\item{BXhat}{Estimated block probability matrix. Should be a \code{k} by \code{k} matrix where \code{k} is the number of blocks.}

\item{cov}{A vector specifying possible value that each covariate could take. For example, if two binary covariates, then \code{cov <- c(2, 2)}.}

\item{covariates}{Observed covariates. Should be a \code{n} by \code{c} matrix or dataframe where \code{n} is the number of nodes and \code{c} is the number of covariates.}

\item{clusters_cov}{An \code{n}-vecotr indicating the block label of each nodes with the effect of covariates where \code{n} is the number nodes.}
}
\value{
A list containing all estimated beta. The length of the list equals to the number of covariates. Using \code{sapply(betahats, sum)} to get the estimate of each beta.
}
\description{
Estimate beta (the effect of covariates) using weighted procedure.
}
\seealso{
\code{\link{grdpg}}
}
\author{
Cong Mu \email{placid8889@gmail.com}
}
