% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateA.R
\name{generateA}
\alias{generateA}
\title{Generate Adjacency Matrix}
\usage{
generateA(n, P, directed = FALSE, type = "bernoulli", seed = 2019)
}
\arguments{
\item{n}{Number of nodes.}

\item{P}{Probability matrix. Shoud be an \code{n} by \code{n} matrix.}

\item{directed}{Whether it is directed. \code{FALSE} by default (undirected).}

\item{type}{Type of network links. 'bernoulli' by default.}

\item{seed}{Random seed for reproductivity. 2019 by default.}
}
\value{
An \code{n} by \code{n} matrix where \code{n} is the number of nodes.
}
\description{
Generate adjacency matrix from probability matrix.
}
\examples{
n <- 4
P <- matrix(runif(n*n), nrow = n)
P[lower.tri(P)] <- t(P)[lower.tri(t(P))]
generateA(n, P)

}
\seealso{
\code{\link{grdpg}}
}
\author{
Cong Mu \email{placid8889@gmail.com}
}
