% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateB.R
\name{generateB}
\alias{generateB}
\title{Generate Block Probability Matrix}
\usage{
generateB(latent, K, d, addCovariates, ...)
}
\arguments{
\item{latent}{True latent position. Suould be a \code{d} by \code{K} matrix where \code{d} is the dimension of latent position and \code{K} is the number of blocks.}

\item{K}{Number of blocks.}

\item{d}{Dimension of latent position.}

\item{addCovariates}{Whether add the effect of covariates. \code{FALSE} suggests without the effect of covariates.}

\item{...}{If \code{addCovariates} is \code{TRUE}, then pass two additional parameters \code{cov} (a vector specifying possible value that each covariate could take) and \code{beta} (effect of covariates).}
}
\value{
A \code{K*prod(cov)} by \code{K*prod(cov)} matrix (assuming discrete covariates) where \code{K} is the number of blocks and \code{cov} is a vector specifying possible value that each covariate could take (1 if without effect of covariates).
}
\description{
Generate block probability matrix from latent position (with the effect of covariates).
}
\examples{
## Without covariates
addCovariates <- FALSE
K <- 2                                           # Number of blocks
d <- 2                                           # Dimension of latent position
latent <- cbind(c(0.63, -0.14), c(0.69, 0.13))   # Latent position

B <- generateB(latent, K, d, addCovariates)


## With covariates
# Hyperparameter
addCovariates <- TRUE
K <- 2                        # Number of blocks
d <- 1                        # Dimension of latent position
latent <- cbind(0.1, 0.7)     # Latent position

# One binary covariate
beta <- 0.3           # If two covariates: c(0,1, 0.3), etc
cov <- 2              # If two binary: c(2, 2), etc

B <- generateB(latent, K, d, addCovariates, cov, beta)

}
\seealso{
\code{\link{grdpg}}
}
\author{
Cong Mu \email{placid8889@gmail.com}
}
