% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateP.R
\name{generateP}
\alias{generateP}
\title{Generate Probability Matrix}
\usage{
generateP(latent, d, block_size, addCovariates, ...)
}
\arguments{
\item{latent}{True latent position. Suould be a \code{d} by \code{K} matrix where \code{d} is the dimension of latent position and \code{K} is the number of blocks.}

\item{d}{Dimension of latent position.}

\item{block_size}{Size of each block.}

\item{addCovariates}{Whether add the effect of covariates. \code{FALSE} suggests without the effect of covariates.}

\item{...}{If \code{addCovariates} is \code{TRUE}, then pass two additional parameters \code{covariates} (observed covariates, \code{n} by \code{c} matrix or dataframe where \code{n} is the number of nodes and \code{c} is the number of covariates) and \code{beta} (effect of covariates).}
}
\value{
A \code{n} by \code{n} matrix where \code{n} is the number of nodes.
}
\description{
Generate probability matrix from latent position (with the effect of covariates).
}
\examples{
## Without covariates
# Hyperparameter
addCovariates <- FALSE
n <- 2000                                       # Number of nodes
K <- 2                                          # Number of blocks
d <- 2                                          # Dimension of latent position
latent <- cbind(c(0.63, -0.14), c(0.69, 0.13))  # Latent position

# Size of each block is balanced
pi <- rep(1/K, K)
block_size <- round(pi * n)

P <- generateP(latent, d, block_size, addCovariates)


## With covariates
# Hyperparameter
addCovariates <- TRUE
n <- 2000                     # Number of nodes
K <- 2                        # Number of blocks
d <- 1                        # Dimension of latent position
latent <- cbind(0.1, 0.7)     # Latent position

# Size of each block is balanced
pi <- rep(1/K, K)
block_size <- round(pi * n)

# One binary covariate
beta <- 0.3           # If two covariates: c(0,1, 0.3), etc
cov <- 2              # If two binary: c(2, 2), etc
covariates <- matrix(rep(1:cov, each = 500, times = K))

P <- generateP(latent, d, block_size, addCovariates, covariates, beta)

}
\seealso{
\code{\link{grdpg}}
}
\author{
Cong Mu \email{placid8889@gmail.com}
}
