% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAwithoutCovariates.R
\name{getAwithoutCovariates}
\alias{getAwithoutCovariates}
\title{Get the Adjacency Matrix without Covariates}
\usage{
getAwithoutCovariates(A, betahat, covariates)
}
\arguments{
\item{A}{Adjacency matrix with the effect of covariates. Should be an \code{n} by \code{n} matrix where \code{n} is the number of nodes.}

\item{betahat}{Estimated beta. Should be a \code{c}-vector where \code{c} is the number of covariates.}

\item{covariates}{Observed covariates. Should be a \code{n} by \code{c} matrix or dataframe where \code{n} is the number of nodes and \code{c} is the number of covariates.}
}
\value{
An adjacency matrix removing the effect of covariates.
}
\description{
Get the adjacency matrix without the effect of covariates.
}
\seealso{
\code{\link{grdpg}}
}
\author{
Cong Mu \email{placid8889@gmail.com}
}
