% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBlockCovariates.R
\name{getBlockCovariates}
\alias{getBlockCovariates}
\title{Get the Covariates for Each Block}
\usage{
getBlockCovariates(covariates, clusters_cov)
}
\arguments{
\item{covariates}{Observed covariates. Should be an \code{n} by \code{c} matrix or dataframe where \code{n} is the number of nodes and \code{c} is the number of covariates.}

\item{clusters_cov}{Block label for each nodes with the effect of covariates. Should be a \code{n}-vector with \code{k} unique values where \code{n} is the number of nodes and \code{k} is the number of blocks with the effect of covariates.}
}
\value{
A \code{k} by \code{c} matrix where \code{k} is the number of blocks with the effect of covariates, \code{c} is the number of covariates and entry \code{[i,j]} is the covariate \code{j} for block \code{i}.
}
\description{
Get the covariates for each block based on the frequency of each covariate within each block.
}
\seealso{
\code{\link{grdpg}}
}
\author{
Cong Mu \email{placid8889@gmail.com}
}
