% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClusters.R
\name{getClusters}
\alias{getClusters}
\title{Get the Block Label of Each Nodes}
\usage{
getClusters(block_assignment_probs)
}
\arguments{
\item{block_assignment_probs}{Estimated block assignment probability. Should be an \code{n} by \code{k} matrix or dataframe where \code{n} is the number of nodes and \code{k} is the number of blocks.}
}
\value{
A vector with block labels of each nodes.
}
\description{
Get the block label of each nodes according to the block assignment probability.
}
\seealso{
\code{\link{grdpg}}
}
\author{
Cong Mu \email{placid8889@gmail.com}
}
