% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIpq.R
\name{getIpq}
\alias{getIpq}
\title{Get \code{Ipq} Matrix for GRDPG}
\usage{
getIpq(s, d)
}
\arguments{
\item{s}{Eigenvalues of a matrix.}

\item{d}{Embeded dimension.}
}
\value{
A \code{d} by \code{d} diagonal matrix where \code{d} is the embeded dimension with 1 and -1 on the diagonal.
}
\description{
Construct \code{Ipq} matrix for Generalized Random Dot Product Graph.
}
\references{
Rubin-Delanchy, P., Priebe, C. E., Tang, M., & Cape, J. (2017). A statistical interpretation of spectral embedding: the generalised random dot product graph. \emph{arXiv preprint \href{https://arxiv.org/abs/1709.05506}{arXiv:1709.05506}}.
}
\seealso{
\code{\link{grdpg}}
}
\author{
Cong Mu \email{placid8889@gmail.com}
}
