% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRDPGwithCovariates.R
\name{GRDPGwithCovariates}
\alias{GRDPGwithCovariates}
\title{(Generalized) Random Dot Product Graph with Covariates}
\usage{
GRDPGwithCovariates(A, covariates, link = "identity",
  clusterMethod = "GMM", G = 1:9, dmax = 10, dhat = NULL,
  maxit = 1000, work = 12, tol = 1e-05, check = "BF",
  postAnalysis = TRUE, plot = TRUE, ...)
}
\arguments{
\item{A}{An adjacency matrix.}

\item{covariates}{Observed covariates. Should be a \code{n} by \code{c} matrix or dataframe where \code{n} is the number of nodes and \code{c} is the number of covariates.}

\item{link}{Link function. Could be 'identity' (by default) or 'logit'.}

\item{clusterMethod}{Method to cluster the estimated latent position. Could be 'GMM' (by default, Gaussian Mixture Model) or 'kmeans'.}

\item{G}{\code{G} for \link[mclust]{Mclust} if \code{clusterMethod=='GMM'} or \code{centers} for \link[stats]{kmeans} if \code{clusterMethod=='kmeans'}. \code{G = 1:9} by default.}

\item{dmax}{Maximal embeded dimension. 10 by default.}

\item{dhat}{Embeded dimension. \code{NULL} by default. If \code{NULL}, will be chosen by \href{http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.90.3768&rep=rep1&type=pdf}{profile likelihood}.}

\item{maxit}{Maximum number of iterations for \code{\link[irlba]{irlba}}.}

\item{work}{Working subspace dimension for \code{\link[irlba]{irlba}}.}

\item{tol}{Stopping tolerance for \code{\link[irlba]{irlba}}.}

\item{check}{Method to check probability matrix. Could be 'BF' (by default, see \link{BFcheck}) or 'Remove' (see \link{Removecheck}).}

\item{postAnalysis}{Whether to do some post analysis such as removing the effect of covariates. \code{TRUE} by default.}

\item{plot}{Whether to show scree plot and latent position. \code{TRUE} by default.}

\item{...}{Additional parameters.}
}
\value{
A list containing the following:
\describe{
\item{\code{Xhat}}{An \code{n} by \code{d} matrix indicating the estimated latent positions, where \code{n} is the number nodes and \code{d} is the embeded dimension.}
\item{\code{betahat_simple}}{A \code{c}-vector indicating the effect of covariates where \code{c} is the number of covariates using simple procedure.}
\item{\code{betahat_weighted}}{A \code{c}-vector indicating the effect of covariates where \code{c} is the number of covariates using weighted procedure.}
\item{\code{BXhat}}{A matrix indicating the estimated block probability matrix.}
\item{\code{clusters_cov}}{An \code{n}-vecotr indicating the block label of each nodes with the effect of covariates where \code{n} is the number nodes.}
\item{\code{Ipq}}{\code{Ipq} matrix for (G)RDPG, see \link{getIpq}.}
\item{\code{iter}}{The number of Lanczos iterations carried out. See \link[irlba]{irlba}.}
\item{\code{mprod}}{The total number of matrix vector products carried out. See \link[irlba]{irlba}.}
\item{\code{Xhatprime_simple}}{If \code{postAnalysis==TRUE}, estimated latent positions after removing the effect of covariates associated with \code{betahat_simple}.}
\item{\code{clusters_simple}}{If \code{postAnalysis==TRUE}, block label of each nodes without the effect of covariates associated with \code{betahat_simple}.}
\item{\code{iterprime_simple}}{If \code{postAnalysis==TRUE}, the number of Lanczos iterations carried out after removing the effect of covariates associated with \code{betahat_simple}. See \link[irlba]{irlba}.}
\item{\code{mprodprime_simple}}{If \code{postAnalysis==TRUE}, the total number of matrix vector products carried out after removing the effect of covariates associated with \code{betahat_simple}. See \link[irlba]{irlba}.}
\item{\code{Xhatprime_weighted}}{If \code{postAnalysis==TRUE}, estimated latent positions after removing the effect of covariates associated with \code{betahat_weighted}.}
\item{\code{clusters_weighted}}{If \code{postAnalysis==TRUE}, block label of each nodes without the effect of covariates associated with \code{betahat_weighted}.}
\item{\code{iterprime_weighted}}{If \code{postAnalysis==TRUE}, the number of Lanczos iterations carried out after removing the effect of covariates associated with \code{betahat_weighted}. See \link[irlba]{irlba}.}
\item{\code{mprodprime_weighted}}{If \code{postAnalysis==TRUE}, the total number of matrix vector products carried out after removing the effect of covariates associated with \code{betahat_weighted}. See \link[irlba]{irlba}.}
\item{\code{pp1}}{Screeplot with covariates.}
\item{\code{pp2}}{Latent position in 2D.}
\item{\code{pp3_simple}}{If \code{postAnalysis==TRUE}, screeplot without covariates associated with \code{betahat_simple}.}
\item{\code{pp3_weighted}}{If \code{postAnalysis==TRUE}, screeplot without covariates associated with \code{betahat_weighted}.}
}
}
\description{
Estimate (Generalized) Random Dot Product Graph (GRDPG) with effect of covariates.
}
\examples{
## Parameters
addCovariates <- TRUE
n <- 2000                  # Number of nodes
K <- 2                     # Number of blocks
d <- 1                     # Dimension of latent position
latent <- cbind(0.1, 0.7)  # Latent position

# Balanced case
pi <- rep(1/K, K)
block_size <- round(pi * n)

# Block label without effect of covariates
blocks <- c()
for (k in 1:length(block_size)) {
  blocks <- c(blocks, rep(k, block_size[k]))
}

## One binary covariate
beta <- 0.3
cov <- 2
pi_cov <- rep(1/(K*prod(cov)), K*prod(cov))
block_size_cov <- round(pi_cov * n)
covariates <- matrix(rep(1:cov, each = 500, times = K))

# Block label with effect of covariates
blocks_cov <- c()
for (k in 1:length(block_size_cov)) {
  blocks_cov <- c(blocks_cov, rep(k, block_size_cov[k]))
}

## Generate network (adjacency matrix)
B <- generateB(latent, K, d, addCovariates, cov, beta)
P <- generateP(latent, d, block_size, addCovariates, covariates, beta)
A <- generateA(n, P)

## Estimation
result <- GRDPGwithCovariates(A, covariates)

## Evaluation
print(adjustedRandIndex(blocks_cov, result$clusters_cov))          # ARI with covariates
print(adjustedRandIndex(blocks, result$clusters_simple))           # ARI without covariates associated with beta_simple
print(adjustedRandIndex(blocks, result$clusters_weighted))         # ARI without covariates associated with beta_weighted
print(beta)                              # True beta
print(result$betahat_simple)             # Estimated beta using simple procedure
print(result$betahat_weighted)           # Estimated beta using weighted procedure
print(B)                                 # True B matrix
print(result$BXhat)                      # Estimated B matrix

## Visualization
pp2 <- plotLatentPosition(result$Xhat, blocks, withCovariates = TRUE, dhat = ncol(result$Xhat), covariates)
pp4_simple <- plotLatentPosition(result$Xhatprime_simple, blocks, withCovariates = FALSE, latent, K, d)
multiplot(result$pp1, result$pp3_simple, pp2[[1]], pp4_simple, cols = 2)

pp4_weighted <- plotLatentPosition(as.matrix(result$Xhatprime_weighted[,nrow(latent)]), blocks, withCovariates = FALSE, latent, K, d)
multiplot(result$pp1, result$pp3_weighted, pp2[[1]], pp4_weighted, cols = 2)

}
\seealso{
\code{\link{grdpg}}
}
\author{
Cong Mu \email{placid8889@gmail.com}
}
