% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRDPGwithoutCovariates.R
\name{GRDPGwithoutCovariates}
\alias{GRDPGwithoutCovariates}
\title{(Generalized) Random Dot Product Graph without Covariates}
\usage{
GRDPGwithoutCovariates(A, link = "identity", clusterMethod = "GMM",
  G = 1:9, dmax = 10, dhat = NULL, maxit = 1000, work = 12,
  tol = 1e-05, check = "BF", plot = TRUE, ...)
}
\arguments{
\item{A}{An adjacency matrix.}

\item{link}{Link function. Could be 'identity' (by default) or 'logit'.}

\item{clusterMethod}{Method to cluster the estimated latent position. Could be 'GMM' (by default, Gaussian Mixture Model) or 'kmeans'.}

\item{G}{\code{G} for \link[mclust]{Mclust} if \code{clusterMethod=='GMM'} or \code{centers} for \link[stats]{kmeans} if \code{clusterMethod=='kmeans'}. \code{G = 1:9} by default.}

\item{dmax}{Maximal embeded dimension. 10 by default.}

\item{dhat}{Embeded dimension. \code{NULL} by default. If \code{NULL}, will be chosen by \href{http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.90.3768&rep=rep1&type=pdf}{profile likelihood}.}

\item{maxit}{Maximum number of iterations for \code{\link[irlba]{irlba}}.}

\item{work}{Working subspace dimension for \code{\link[irlba]{irlba}}.}

\item{tol}{Stopping tolerance for \code{\link[irlba]{irlba}}.}

\item{check}{Method to check probability matrix. Could be 'BF' (by default, see \link{BFcheck}) or 'Remove' (see \link{Removecheck}).}

\item{plot}{Whether to show scree plot and latent position. \code{TRUE} by default.}

\item{...}{Additional parameters.}
}
\value{
A list containing the following:
\describe{
\item{\code{Xhat}}{An \code{n} by \code{d} matrix indicating the estimated latent positions, where \code{n} is the number nodes and \code{d} is the embeded dimension.}
\item{\code{BXhat}}{A matrix indicating the estimated block probability matrix.}
\item{\code{clusters_cov}}{An \code{n}-vecotr indicating the block label of each nodes with the effect of covariates where \code{n} is the number nodes.}
\item{\code{Ipq}}{\code{Ipq} matrix for (G)RDPG, see \link{getIpq}.}
\item{\code{pp1}}{Screeplot with covariates.}
}
}
\description{
Estimate (Generalized) Random Dot Product Graph (GRDPG) without effect of covariates.
}
\examples{
## Parameters
addCovariates <- FALSE
n <- 2000                  # Number of nodes
K <- 2                     # Number of blocks
d <- 1                     # Dimension of latent position
latent <- cbind(0.1, 0.7)  # Latent position

# Balanced case
pi <- rep(1/K, K)
block_size <- round(pi * n)

# Block label without effect of covariates
blocks <- c()
for (k in 1:length(block_size)) {
  blocks <- c(blocks, rep(k, block_size[k]))
}

## Generate network (adjacency matrix)
B <- generateB(latent, K, d, addCovariates)
P <- generateP(latent, d, block_size, addCovariates)
A <- generateA(n, P)

## Estimation
result <- GRDPGwithoutCovariates(A)

## Evaluation
print(adjustedRandIndex(blocks, result$clusters))            # ARI without covariates
print(B)                        # True B matrix
print(result$BXhat)             # Estimated B matrix

## Visualization
plotLatentPosition(matrix(result$Xhat[,1]), blocks, FALSE, latent, K, d)

}
\seealso{
\code{\link{grdpg}}
}
\author{
Cong Mu \email{placid8889@gmail.com}
}
