% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Logit Function}
\usage{
logit(p)
}
\arguments{
\item{p}{Input of the logit function. Could be a number, vector or matrix.}
}
\value{
Logit of the input \code{p}. If \code{p} is a vector or matrix, it will return the element-wise logit of \code{p}.
}
\description{
Logit function.
}
\details{
\href{https://en.wikipedia.org/wiki/Logit}{Logit function}
}
\examples{
P <- matrix(runif(4), nrow = 2)
logit(P)

}
\seealso{
\code{\link{grdpg}}
}
\author{
Cong Mu \email{placid8889@gmail.com}
}
