% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitderivative.R
\name{logitderivative}
\alias{logitderivative}
\title{Derivative of logit Function}
\usage{
logitderivative(x)
}
\arguments{
\item{x}{Input of the derivative of logit function. Could be a number, vector or matrix.}
}
\value{
Derivative of logit of the input \code{x}. If \code{x} is a vector or matrix, it will return the element-wise derivative of logit of \code{x}.
}
\description{
Derivative of logit Function.
}
\examples{
X <- matrix(runif(4), nrow = 2)
logitderivative(X)

}
\seealso{
\code{\link{grdpg}}
}
\author{
Cong Mu \email{placid8889@gmail.com}
}
