% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple Plots on One Page for \code{ggplot2}}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{A list of \code{ggplot2} objects.}

\item{plotlist}{A list of \code{ggplot2} objects. \code{NULL} by default.}

\item{cols}{Number of columns in layout. 1 by default.}

\item{layout}{A matrix specifying the layout. \code{NULL} by default. If present, \code{cols} is ignored. For example, if the layout is \code{matrix(c(1,2,3,3), nrow=2, byrow=TRUE)}, then plot 1 will go in the upper left, 2 will go in the upper right, and 3 will go all the way across the bottom.}
}
\description{
Show multiple plots on one page for \code{ggplot2}.
}
\examples{
require(ggplot2)
p1 <- ggplot(economics, aes(date, unemploy)) + geom_line()
p2 <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
multiplot(p1, p2)

}
\references{
\href{http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/}{Cookbook for R}
}
\seealso{
\code{\link{grdpg}}
}
