% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLatentPosition.R
\name{plotLatentPosition}
\alias{plotLatentPosition}
\title{Plot Latent Position}
\usage{
plotLatentPosition(Xhat, blocks = NULL, withCovariates, ...)
}
\arguments{
\item{Xhat}{Estimated latent position. Should be a \code{n} by \code{d} matrix where \code{n} is the number of nodes and \code{d} is the embeded dimension.}

\item{blocks}{Block label of each nodes. Should be a \code{n}-vector where \code{n} is the number of nodes. \code{NULL} by default.}

\item{withCovariates}{Whether with effect of covariates. \code{FALSE} suggests without effect of covariates.}

\item{...}{Additional parameters based on the value of \code{withCovariates}
\itemize{
\item{With covariates}
\describe{
\item{\code{dhat}}{Embeded dimension}
\item{\code{covariates}}{Observed covariates}
}
\item{Without covariates}
\describe{
\item{\code{latent}}{True latent position}
\item{\code{K}}{Number of blocks}
\item{\code{d}}{Dimension of latent position}
}
}}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot latent position (with the effect of covariates).
}
\seealso{
\code{\link{grdpg}}
}
\author{
Cong Mu \email{placid8889@gmail.com}
}
