% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Removecheck.R
\name{Removecheck}
\alias{Removecheck}
\title{Remove Check of Probability Matrix}
\usage{
Removecheck(P)
}
\arguments{
\item{P}{Estimated probability matrix.}
}
\value{
A matrix of which all entries are between 0 and 1.
}
\description{
Check whether all entries of the estimated probabiltiy matrix are between 0 and 1.
Remove all rows and columns that contain values that are not greater than 0 and not less than 1.
}
\examples{
P <- matrix(runif(100,-1,2), nrow = 10)
Removecheck(P)

}
\seealso{
\code{\link{grdpg}}
}
\author{
Cong Mu \email{placid8889@gmail.com}
}
