% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Rotation of Estimated Latent Position}
\usage{
rotate(Xhat, latent, K)
}
\arguments{
\item{Xhat}{Estimated latent position. Should be an \code{n} by \code{d} matrix where \code{n} is the number of nodes and \code{d} is the embeded dimension.}

\item{latent}{True latent position. Suould be a \code{d} by \code{K} matrix where \code{d} is the dimension of latent position and \code{K} is the number of blocks.}

\item{K}{Number of blocks.}
}
\value{
Rotated estimated latent position that could be compared with the truth.
}
\description{
Rotate estimated latent position to compare with the truth.
}
\seealso{
\code{\link{grdpg}}
}
\author{
Cong Mu \email{placid8889@gmail.com}
}
