% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scree.R
\name{scree}
\alias{scree}
\title{Scree Plot of Eigenvalues (Singular Values)}
\usage{
scree(s, title = "Screeplot", xlab = "Rank in Magnitude",
  ylab = "Eigenvalue in Magnitude")
}
\arguments{
\item{s}{Eigenvalues (singular values) of a matrix.}

\item{title}{Title of the plot. 'Screeplot' by default.}

\item{xlab}{Label of x-axis. 'Rank in Magnitude' by default.}

\item{ylab}{Label of y-axis. 'Eigenvalue in Magnitude' by default.}
}
\value{
A \code{ggplo2} object.
}
\description{
Scree plot of the eigenvalues (singular values) of a matrix.
}
\examples{
X <- matrix(runif(100), nrow = 10)
X <- X + t(X)
dicomp <- svd(X)
dicomp2 <- eigen(X)
scree(dicomp$d, xlab = 'Rank', ylab = 'Singular Value')
scree(dicomp2$values)

}
\seealso{
\code{\link{grdpg}}
}
\author{
Cong Mu \email{placid8889@gmail.com}
}
