% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid Function}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{Input of the sigmoid function. Could be a number, vector or matrix.}
}
\value{
Sigmoid of the input \code{x}. If \code{x} is a vector or matrix, it will return the element-wise sigmoid of \code{x}.
}
\description{
Sigmoid function.
}
\details{
\href{https://en.wikipedia.org/wiki/Sigmoid_function}{Sigmoid function}
}
\examples{
X <- matrix(runif(4), nrow = 2)
sigmoid(X)

}
\seealso{
\code{\link{grdpg}}
}
\author{
Cong Mu \email{placid8889@gmail.com}
}
