% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{(Generalized) Random Dot Product Graph with Covariates}
\usage{
simulation(n, K, d, latent, block_size, beta, cov, block_size_cov,
  covariates, link = "identity", clusterMethod = "GMM", G = 1:9,
  dmax = 10, dhat = NULL, maxit = 1000, work = 12, tol = 1e-05,
  check = "BF", sd = TRUE, rho = 1, postAnalysis = TRUE,
  plot = TRUE, seed = 2019, ...)
}
\arguments{
\item{n}{Number of nodes.}

\item{K}{Number of blocks.}

\item{d}{Dimension of latent position.}

\item{latent}{True latent position. Suould be a \code{d} by \code{K} matrix where \code{d} is the dimension of latent position and \code{K} is the number of blocks.}

\item{block_size}{The size of each block without the effect of covariates.}

\item{beta}{The effect of covariates. Should be a \code{c}-vector where \code{c} is the number of covariates.}

\item{cov}{Possible values of each covariate could take (assuming discrete here). Should be a \code{c}-vector where \code{c} is the number of covariates. For example, if two binary covariates, then \code{cov <- c(2, 2)}.}

\item{block_size_cov}{The size of each block with the effect of covariates.}

\item{covariates}{Observed covariates. Should be a \code{n} by \code{c} matrix or dataframe where \code{n} is the number of nodes and \code{c} is the number of covariates.}

\item{link}{Link function. Could be 'identity' (by default) or 'logit'.}

\item{clusterMethod}{Method to cluster the estimated latent position. Could be 'GMM' (by default, Gaussian Mixture Model) or 'kmeans'.}

\item{G}{\code{G} for \link[mclust]{Mclust} if \code{clusterMethod=='GMM'} or \code{centers} for \link[stats]{kmeans} if \code{clusterMethod=='kmeans'}. \code{G = 1:9} by default.}

\item{dmax}{Maximal embeded dimension. 10 by default.}

\item{dhat}{Embeded dimension. \code{NULL} by default. If \code{NULL}, will be chosen by \href{http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.90.3768&rep=rep1&type=pdf}{profile likelihood}.}

\item{maxit}{Maximum number of iterations for \code{\link[irlba]{irlba}}.}

\item{work}{Working subspace dimension for \code{\link[irlba]{irlba}}.}

\item{tol}{Stopping tolerance for \code{\link[irlba]{irlba}}.}

\item{check}{Method to check probability matrix. Could be 'BF' (by default, see \link{BFcheck}) or 'Remove' (see \link{Removecheck}).}

\item{sd}{Whether to compute standard errors of the estimate of beta. \code{TRUE} by default.}

\item{rho}{Sparsity coefficient. Coule be \code{1} (by default) or \code{0}.}

\item{postAnalysis}{Whether to do some post analysis such as removing the effect of covariates. \code{TRUE} by default.}

\item{plot}{Whether to show scree plot and latent position. \code{TRUE} by default.}

\item{seed}{Random seed for reproductivity. 2019 by default.}

\item{...}{Additional parameters.}
}
\description{
Simulate and estimate (Generalized) Random Dot Product Graph (GRDPG) with effect of covariates.
}
\examples{
## Parameters
addCovariates <- TRUE
n <- 2000                  # Number of nodes
K <- 2                     # Number of blocks
d <- 1                     # Dimension of latent position
latent <- cbind(0.1, 0.7)  # Latent position

# Balanced case
pi <- rep(1/K, K)
block_size <- round(pi * n)

## One binary covariate
beta <- 0.3
cov <- 2
pi_cov <- rep(1/(K*prod(cov)), K*prod(cov))
block_size_cov <- round(pi_cov * n)
covariates <- matrix(rep(1:cov, each = 500, times = K))

## Simulation
simulation(n, K, d, latent, block_size, beta, cov, block_size_cov, covariates)

}
\seealso{
\code{\link{grdpg}}
}
\author{
Cong Mu \email{placid8889@gmail.com}
}
