% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectralEmbedding.R
\name{SpectralEmbedding}
\alias{SpectralEmbedding}
\title{Spectral Embedding}
\usage{
SpectralEmbedding(g, k = NULL, edge.attr = NULL, maxit = 1000,
  work = 12, tol = 1e-05, ...)
}
\arguments{
\item{g}{An \code{\link[igraph]{igraph}} object or an \code{n} by \code{n} adjacency matrix with \code{n} vertices.}

\item{k}{Embeded dimension. \code{NULL} by default. \code{k} should be less than the number of vertices in \code{g}. If \code{k==NULL}, defaults to \code{floor(log2(nvertices))}.}

\item{edge.attr}{If \code{g} is a \code{\link[igraph]{igraph}}, the name of the attribute to use for weights. Defaults to \code{NULL}, which assumes the graph is binary.
\itemize{
\item{\code{is.null(edge.attr)} Assumes unweighted.}
\item{\code{is.character(edge.attr)} Assumes weighted with weights given by \code{edge.attr}.}
}}

\item{maxit}{Maximum number of iterations for \code{\link[irlba]{irlba}}.}

\item{work}{Working subspace dimension for \code{\link[irlba]{irlba}}.}

\item{tol}{Stopping tolerance for \code{\link[irlba]{irlba}}.}

\item{...}{Other parameters for \code{\link[irlba]{irlba}}.}
}
\value{
A list containing the following:
\describe{
\item{\code{X}}{An \code{n} by \code{k} matrix indicating the estimated latent positions, where \code{n} is the number of vertices of \code{g}.}
\item{\code{Y}}{\code{NULL} if \code{g} is undirected. If \code{g} is directed, \code{Y} is a \code{n} by \code{k} matrix indicating the second half of the latent positions.}
\item{\code{D}}{The eigenvalues (for undirected graphs) or the singular values (for directed graphs) associated with the latent positions.}
\item{\code{iter}}{The number of Lanczos iterations carried out. See \link[irlba]{irlba}.}
\item{\code{mprod}}{The total number of matrix vector products carried out. See \link[irlba]{irlba}.}
}
}
\description{
Spectral decomposition of the adjacency/Laplacian matrices of graphs.
}
\references{
\code{\link[graphstats]{gs.embed}}, \code{\link[graphstats]{gs.embed.ase}}, \code{\link[graphstats]{gs.embed.lse}}, \code{\link[irlba]{irlba}}
}
\seealso{
\code{\link{grdpg}}
}
