% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scree.R
\name{scree}
\alias{scree}
\title{Scree Plot of Singular Values}
\usage{
scree(s, title = "Screeplot", xlab = "Rank", ylab = "Singular Value")
}
\arguments{
\item{s}{Singular values of a matrix.}

\item{title}{Title of the plot. 'Screeplot' by default.}

\item{xlab}{Label of x-axis. 'Rank' by default.}

\item{ylab}{Label of y-axis. 'Singular Value' by default.}
}
\value{
A \code{ggplo2} object.
}
\description{
Scree plot of the singular values of a matrix.
}
\examples{
X <- matrix(runif(100), nrow = 10)
dicomp <- svd(X)
scree(dicomp$d)

}
\seealso{
\code{\link{grdpg}}
}
\author{
Cong Mu \email{placid8889@gmail.com}
}
